+++ R/api-taxonomy.R
@@ -11,8 +11,11 @@
 ##' @importFrom assertthat is.flag
 ##' @importFrom assertthat assert_that
 ## Information about an OpenTree Taxonomy (OTT) taxon
-.taxonomy_taxon <- function(ott_id=NULL, include_lineage = FALSE,
-                            list_terminal_descendants = FALSE, ...) {
+.taxonomy_taxon_info <- function(ott_id=NULL,
+                                 include_children = FALSE,
+                                 include_lineage = FALSE,
+                                 include_terminal_descendants = FALSE,
+                                  ...) {
     if (is.null(ott_id)) {
         stop("Must supply an \'ott_id\' argument")
     } else if (length(ott_id) > 1) {
@@ -20,12 +23,14 @@
     } else if (!check_numeric(ott_id)) {
         stop("Argument \'ott_id\' must look like a number.")
     }
+    assertthat::assert_that(assertthat::is.flag(include_children))
     assertthat::assert_that(assertthat::is.flag(include_lineage))
-    assertthat::assert_that(assertthat::is.flag(list_terminal_descendants))
+    assertthat::assert_that(assertthat::is.flag(include_terminal_descendants))
     q <- list(ott_id=jsonlite::unbox(ott_id),
+              include_children = jsonlite::unbox(include_children),
               include_lineage = jsonlite::unbox(include_lineage),
-              list_terminal_descendants = jsonlite::unbox(list_terminal_descendants))
-    res <- otl_POST(path="/taxonomy/taxon", body=q, ...)
+              include_terminal_descendants = jsonlite::unbox(include_terminal_descendants))
+    res <- otl_POST(path="/taxonomy/taxon_info", body=q, ...)
     res
 }
 
@@ -33,7 +38,7 @@
 ##' @importFrom jsonlite unbox
 ##' @importFrom httr content
 ## Get a subtree from the OpenTree Taxonomy (OTT) taxonomic tree
-.taxonomy_subtree <- function(ott_id=NULL, ...) {
+.taxonomy_subtree <- function(ott_id=NULL, label_format=NULL, ...) {
     if (is.null(ott_id)) {
         stop("Must supply an \'ott_id\' argument")
     } else if (length(ott_id) > 1) {
@@ -42,20 +47,27 @@
         stop("Argument \'ott_id\' must look like a number.")
     }
     q <- list(ott_id=jsonlite::unbox(ott_id))
+    if (!is.null(label_format)) {
+        if (!check_label_format(label_format)) {
+            stop(sQuote("label_format"), " must be one of: ", sQuote("name"), ", ",
+                 sQuote("id"), ", or ", sQuote("name_and_id"))
+        }
+        q$label_format <- jsonlite::unbox(label_format)
+    }
     res <- otl_POST(path="/taxonomy/subtree", body=q, ...)
     res
 }
 
 
 ##' @importFrom httr content
-## Get the least inclusive common ancestor (LICA) from nodes in the OpenTree Taxonomy (OTT)
-.taxonomy_lica <- function (ott_ids = NULL, ...) {
+## Get the most recent common ancestor (MRCA) from nodes in the OpenTree Taxonomy (OTT)
+.taxonomy_mrca <- function (ott_ids = NULL, ...) {
     if (is.null(ott_ids)) {
         stop("Must supply an \'ott_ids\' argument")
     } else if (!all(sapply(ott_ids, check_numeric))) {
         stop("Argument \'ott_ids\' must look like a number.")
     }
     q <- list(ott_ids=ott_ids)
-    res <- otl_POST(path="/taxonomy/lica", body=q, ...)
+    res <- otl_POST(path="/taxonomy/mrca", body=q, ...)
     res
 }